+++ src/main/front-end/src/components/news-feed-page/filters-button/index.ts
@@ -1,0 +1,1 @@
+export * from "./FiltersDropDownButton";
