+++ src/main/front-end/src/components/news-feed-page/recommended-teacher/index.ts
@@ -1,0 +1,1 @@
+export * from "./RecommendedTeacher";
