+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url = jdbc:postgresql://localhost:5432/postgres
+spring.datasource.url = jdbc:postgresql://localhost:8088/postgres
 #---enter your username and password------------------
 spring.datasource.username = postgres
 spring.datasource.password = password
